requirements.txt

google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openai==1.51.0
ollama==0.4.7
SpeechRecognition==3.10.4
PyAudio==0.2.14
python-dotenv==1.0.1
httplib2==0.22.0
httpx==0.27.0
rich==13.7.1
streamlit==1.38.0
audio-recorder-streamlit==0.0.10
